Implementent foundation of scripting support
Script commands can either be specified on the command line (--do) or
read from a file (--do-file). Every command can either be an assignment
affecting the simstate or call an action.

To make this possible, the simstate is now globally created in __main__.py.
Additionally, The value-parsing code has been made independend of configparser.