Add parameter "inverted_phase_correction"
This is compatibility hack for old cospe data that was range compressed in the wrong way.
The other way to fix this, is to use a negative frequency in the config file for the data.
Negative frequencies are not allowed in the GUI, so this break when the GUI is loaded.
This seems to be the safer way.